requirements.txt
Django==2.2.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
Pillow==6.0.0
pytz==2019.1
qrcode==6.1
six==1.12.0
sqlparse==0.3.0
